create a cache package
